requirements.txt
streamlit
streamlit-audiorecorder
moviepy
pydub
SpeechRecognition
google-generativeai
python-dotenv
